refactor: move onAuthStateChange to hooks
