+++ src/store/hooks.ts
@@ -1,3 +1,5 @@
+import { removeUser, setUser } from './slices/user'
+import { getAuth, onAuthStateChanged } from 'firebase/auth'
 import { useEffect, useState } from 'react'
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux'
 import type { RootState, AppDispatch } from './store'
@@ -7,10 +9,11 @@ export const useAppDispatch = () => useDispatch<AppDispatch>()
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
 
 export const useAuthorization = () => {
-    const { email, token, id } = useAppSelector(state => state.user)
+    const { email, token, id, isAuth } = useAppSelector(state => state.user)
+    useEffect
 
     return {
-        isAuth: !!email,
+        isAuth,
         email,
         token,
         id,
@@ -31,4 +34,28 @@ export function useDebounce<T>(value: T | string, delay: number) {
     }, [value])
 
     return debouncedValue
+}
+
+export const useFirebaseAuth = () => {
+    const auth = getAuth()
+    const dispatch = useAppDispatch()
+    useEffect(() => {
+        const unsubscribe = onAuthStateChanged(auth, user => {
+            if (user) {
+                dispatch(
+                    setUser({
+                        email: user.email,
+                        id: user.uid,
+                    })
+                )
+                localStorage.setItem('isUserSignedIn', 'true')
+            } else {
+                dispatch(removeUser())
+                localStorage.setItem('isUserSignedIn', '')
+                localStorage.setItem('gameName', '')
+                localStorage.setItem('pageNumber', '1')
+            }
+        })
+        return unsubscribe()
+    }, [])
 }
